+++ extensions/iceworks-app/src/extension.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 import { Terminal, window, ViewColumn } from 'vscode';
 import { connectService, getHtmlForWebview } from '@iceworks/vscode-webview/lib/vscode';
 import { getProjectType } from '@iceworks/project-service';
-import { initExtension, Logger } from '@iceworks/common-service';
+import { initExtension, Logger, checkIsAliInternal } from '@iceworks/common-service';
 import { createNpmScriptsTreeProvider } from './views/npmScriptsView';
 import { createNodeDependenciesTreeProvider } from './views/nodeDependenciesView';
 import { createComponentsTreeProvider } from './views/componentsView';
@@ -11,7 +11,9 @@ import { ITerminalMap } from './types';
 import services from './services';
 import { showExtensionsQuickPickCommandId } from './constants';
 import showExtensionsQuickPick from './quickPicks/showExtensionsQuickPick';
+import showDefPublishEnvQuickPick from './quickPicks/showDefPublishEnvQuickPick';
 import createExtensionsStatusBar from './statusBar/createExtensionsStatusBar';
+import i18n from './i18n';
 
 // eslint-disable-next-line
 const { name, version } = require('../package.json');
@@ -22,8 +24,8 @@ export async function activate(context: vscode.ExtensionContext) {
 
   // data collection
   const logger = new Logger(name, globalState);
-  logger.recordDAU();
-  logger.recordActivate(version);
+  logger.recordMainDAU();
+  logger.recordExtensionActivate(version);
 
   // auto set configuration
   initExtension(context);
@@ -34,10 +36,11 @@ export async function activate(context: vscode.ExtensionContext) {
   subscriptions.push(extensionsStatusBar);
   // init webview
   function activeWebview() {
-    const webviewPanel: vscode.WebviewPanel = window.createWebviewPanel('iceworks', '设置 - Iceworks', ViewColumn.One, {
-      enableScripts: true,
-      retainContextWhenHidden: true,
-    });
+    const webviewPanel: vscode.WebviewPanel = window.createWebviewPanel('iceworks', 
+      i18n.format('extension.iceworksApp.extension.title'), ViewColumn.One, {
+        enableScripts: true,
+        retainContextWhenHidden: true,
+      });
     webviewPanel.webview.html = getHtmlForWebview(extensionPath);
     connectService(webviewPanel, context, { services, logger });
   }
@@ -47,7 +50,7 @@ export async function activate(context: vscode.ExtensionContext) {
   }));
 
   if (!rootPath) {
-    vscode.window.showInformationMessage('当前工作区为空，请打开应用或新建应用。');
+    vscode.window.showInformationMessage(i18n.format('extension.iceworksApp.extebsion.emptyWorkplace'));
     vscode.commands.executeCommand('setContext', 'iceworks:isNotTargetProject', true);
     vscode.commands.executeCommand('iceworks-project-creator.start');
     return;
@@ -63,10 +66,19 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.commands.executeCommand('setContext', 'iceworks:isNotTargetProject', true);
     vscode.commands.executeCommand('iceworks-project-creator.start');
   }
+
   const terminals: ITerminalMap = new Map<string, Terminal>();
   // init tree data providers
   createNpmScriptsTreeProvider(context, rootPath, terminals);
   createComponentsTreeProvider(context, rootPath);
   createPagesTreeProvider(context, rootPath);
   createNodeDependenciesTreeProvider(context, rootPath, terminals);
+  // show script icons in editor title menu
+  vscode.commands.executeCommand('setContext', 'iceworks:showScriptIconInEditorTitleMenu', true);
+  const isAliInternal = await checkIsAliInternal();
+  // DEF publish command in editor title
+  vscode.commands.executeCommand('setContext', 'iceworks:isAliInternal', isAliInternal);
+  if (isAliInternal) {
+    context.subscriptions.push(vscode.commands.registerCommand('iceworksApp.DefPublish', () => showDefPublishEnvQuickPick(terminals, rootPath)));
+  }
 }