+++ extensions/iceworks-app/web/src/pages/ConfigHelper/CustomMaterialSource.tsx
@@ -1,6 +1,7 @@
 import React, { useState, Suspense } from 'react';
 import { Icon, List, Button, Dialog, Avatar } from '@alifd/next';
 import { IMaterialSource } from '@iceworks/material-utils';
+import { FormattedMessage, useIntl } from 'react-intl';
 import editIcon from '../../../public/assets/edit.svg';
 import deleteIcon from '../../../public/assets/delete.svg';
 import MaterialSourceForm from './MaterialSourceForm';
@@ -25,6 +26,7 @@ const CustomMaterialSource: React.FC<ICustomMaterialSource> = ({
   onSourceEdit,
   onSourceDelete
 }) => {
+  const intl = useIntl();
   const [visible, setVisible] = useState<boolean>(false);
   const [currentMaterialSource, setCurrentMaterialSource] = useState<IMaterialSource | object>({});
   const [operation, setOperation] = useState<Operation.Create | Operation.Edit>();
@@ -56,18 +58,21 @@ const CustomMaterialSource: React.FC<ICustomMaterialSource> = ({
   }
 
   const onDelete = (materialSource: IMaterialSource) => {
+
     Dialog.confirm({
       title: 'Confirm',
-      content: '是否确定删除该物料源？',
+      content: intl.formatMessage({id:'web.iceworksApp.customMaterialSource.confirmDelete'}),
       onOk: () => onSourceDelete(materialSource)
     });
   }
-  const dialogTitle = `${operation === Operation.Edit ? '编辑' : '新增'}物料源`
+  const dialogTitle = operation === Operation.Edit ? 
+    intl.formatMessage({id:'web.iceworksApp.customMaterialSource.editMaterialSource'}):
+    intl.formatMessage({id:'web.iceworksApp.customMaterialSource.addMaterialSource'});
   return (
     <div className={styles.customMaterialSource}>
       <div className={styles.row}>
-        <span className={styles.label}>自定义物料源</span>
-        <div className={styles.btn}><Button onClick={onAdd}><Icon type="add" />添加</Button></div>
+        <span className={styles.label}><FormattedMessage id='web.iceworksApp.customMaterialSource.customMaterialSource'/></span>
+        <div className={styles.btn}><Button onClick={onAdd}><Icon type="add" /><FormattedMessage id='web.iceworksApp.customMaterialSource.add'/></Button></div>
       </div>
       <div className={styles.sourcesList}>
         <List size="small">
@@ -88,7 +93,7 @@ const CustomMaterialSource: React.FC<ICustomMaterialSource> = ({
           ))}
         </List>
       </div>
-      <Suspense fallback="loading...">
+      <Suspense fallback={intl.formatMessage({id:'web.iceworksApp.customMaterialSource.loading'})}>
         <MaterialSourceForm
           value={currentMaterialSource}
           title={dialogTitle}