+++ extensions/iceworks-app/web/src/pages/ConfigHelper/index.tsx
@@ -1,9 +1,11 @@
 import React, { useState, useEffect } from 'react';
-import { Form, Select, Input, Notification } from '@alifd/next';
+import { Form, Select, Input, Notification, Loading } from '@alifd/next';
 import { debounce } from 'throttle-debounce';
 import { IMaterialSource } from '@iceworks/material-utils';
-import { packageManagers, npmRegistries, AliNpmRegistry, AliPackageManager } from '@/constants';
+import { packageManagers, npmRegistries, AliNpmRegistry, AliPackageManager, urlRegExp } from '@/constants';
 import callService from '@/callService';
+import { useIntl } from 'react-intl'
+import { LocaleProvider} from '../../i18n';
 import CustomMaterialSource from './CustomMaterialSource';
 import styles from './index.module.scss';
 
@@ -16,15 +18,17 @@ const formItemLayout = {
 const CUSTOM_NPM_REGISTRY_FORM_ITEM_KEY = 'customNpmRegistry';
 const CUSTOM_NPM_REGISTRY_SELECT_KEY = 'npm - 自定义镜像源';
 
-const ConfigHelper = () => {
+const ConfigHelper = ( props ) => {
   const [materialSources, setMaterialSources] = useState<IMaterialSource[]>([]);
   const [fields, setFields] = useState<any>({});
+  const [loading, setLoading] = useState<boolean>(false);
+  const intl = useIntl();
 
   const onSourceAdd = async (materialSource: IMaterialSource) => {
     try {
       const newMaterialSources = await callService('material', 'addSource', materialSource);
       setMaterialSources(newMaterialSources);
-      Notification.success({ content: '新增物料源成功' });
+      Notification.success({ content: intl.formatMessage({id:'web.iceworksApp.index.addMaterialSuccess'}) });
     } catch (e) {
       Notification.error({ content: e.message });
     }
@@ -34,7 +38,7 @@ const ConfigHelper = () => {
     try {
       const newMaterialSources = await callService('material', 'updateSource', materialSource, originMaterialSource);
       setMaterialSources(newMaterialSources);
-      Notification.success({ content: '修改物料源成功' });
+      Notification.success({ content:  intl.formatMessage({id:'web.iceworksApp.index.editMaterialSuccess'}) });
     } catch (e) {
       Notification.error({ content: e.message });
     }
@@ -44,7 +48,7 @@ const ConfigHelper = () => {
     try {
       const newMaterialSources = await callService('material', 'removeSource', materialSource.source);
       setMaterialSources(newMaterialSources);
-      Notification.success({ content: '删除物料源成功' });
+      Notification.success({ content:  intl.formatMessage({id:'web.iceworksApp.index.deleteMaterialSuccess'}) });
     } catch (e) {
       Notification.error({ content: e.message });
     }
@@ -57,7 +61,7 @@ const ConfigHelper = () => {
         return;
       }
       if (name === CUSTOM_NPM_REGISTRY_FORM_ITEM_KEY &&
-        !/^(https?:\/\/(([a-zA-Z0-9]+-?)+[a-zA-Z0-9]+\.)+[a-zA-Z]+)(:\d+)?(\/.*)?(\?.*)?(#.*)?$/.test(value)) {
+        !urlRegExp.test(value)) {
         return;
       }
       if (name === CUSTOM_NPM_REGISTRY_FORM_ITEM_KEY) {
@@ -65,7 +69,7 @@ const ConfigHelper = () => {
       } else {
         await callService('common', 'saveDataToSettingJson', name, value);
       }
-      Notification.success({ content: '设置成功' });
+      Notification.success({ content: intl.formatMessage({id:'web.iceworksApp.index.editSettingSuccess'}) });
     } catch (e) {
       Notification.error({ content: e.message });
     }
@@ -74,6 +78,7 @@ const ConfigHelper = () => {
   useEffect(() => {
     async function initFormData() {
       try {
+        setLoading(true);
         const curPackageManager = await callService('common', 'getDataFromSettingJson', 'packageManager');
         let curNpmRegistry = await callService('common', 'getDataFromSettingJson', 'npmRegistry');
         const curMaterialSources = await callService('material', 'getUserSources');
@@ -89,9 +94,12 @@ const ConfigHelper = () => {
           customNpmRegistry = curNpmRegistry;
           curNpmRegistry = CUSTOM_NPM_REGISTRY_SELECT_KEY;
         }
+
         setFields({ packageManager: curPackageManager, npmRegistry: curNpmRegistry, customNpmRegistry });
       } catch (e) {
         Notification.error({ content: e.message });
+      } finally {
+        setLoading(false);
       }
     }
 
@@ -99,36 +107,48 @@ const ConfigHelper = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
   return (
-    <div className={styles.container}>
-      <Form value={fields} {...formItemLayout} labelTextAlign="left" size="medium" onChange={onFormChange}>
-        <FormItem label="npm 包管理工具">
-          <Select name="packageManager" placeholder="请选择 npm 包管理工具" style={{ width: '100%' }}>
-            {packageManagers.map(item => (
-              <Select.Option key={item} value={item}>{item}</Select.Option>
-            ))}
-          </Select>
-        </FormItem>
-        <FormItem label="npm 镜像源">
-          <Select name="npmRegistry" placeholder="请选择 npm 镜像源" style={{ width: '100%' }}>
-            {npmRegistries.map(item => (
-              <Select.Option key={item} value={item}>{item}</Select.Option>
-            ))}
-          </Select>
-        </FormItem>
-        {fields.npmRegistry === CUSTOM_NPM_REGISTRY_SELECT_KEY && (
-          <FormItem label=" " format="url" >
-            <Input name="customNpmRegistry" placeholder="请输入自定义 npm 镜像源" />
-          </FormItem>
-        )}
-      </Form>
-      <CustomMaterialSource
-        sources={materialSources}
-        onSourceAdd={onSourceAdd}
-        onSourceDelete={onMaterialSourceDelete}
-        onSourceEdit={onMaterialSourceEdit}
-      />
-    </div >
+    <>
+      {loading ? <Loading visible={loading} className={styles.loading} /> : (
+        <div className={styles.container}>
+          <Form value={fields} {...formItemLayout} labelTextAlign="left" size="medium" onChange={onFormChange}>
+            <FormItem label={intl.formatMessage({id:'web.iceworksApp.index.npmPackageManager'})}>
+              <Select name="packageManager" placeholder={intl.formatMessage({id:'web.iceworksApp.index.chooseNpmPackageManager'})} style={{ width: '100%' }}>
+                {packageManagers.map(item => (
+                  <Select.Option key={item} value={item}>{item}</Select.Option>
+                ))}
+              </Select>
+            </FormItem>
+            <FormItem label={intl.formatMessage({id:'web.iceworksApp.index.npmRegistry'})}>
+              <Select name="npmRegistry" placeholder={intl.formatMessage({id:'web.iceworksApp.index.chooseNpmRegistry'})} style={{ width: '100%' }}>
+                {npmRegistries.map(item => (
+                  <Select.Option key={item} value={item}>{item}</Select.Option>
+                ))}
+              </Select>
+            </FormItem>
+            {fields.npmRegistry === CUSTOM_NPM_REGISTRY_SELECT_KEY && (
+              <FormItem label=" " format="url" formatMessage={intl.formatMessage({id:'web.iceworksApp.index.formatUrl'})}>
+                <Input name="customNpmRegistry" placeholder={intl.formatMessage({id:'web.iceworksApp.index.customNpmRegistryPlaceHolder'})} />
+              </FormItem>
+            )}
+          </Form>
+          <CustomMaterialSource
+            sources={materialSources}
+            onSourceAdd={onSourceAdd}
+            onSourceDelete={onMaterialSourceDelete}
+            onSourceEdit={onMaterialSourceEdit}
+          />
+        </div>
+      )}
+    </>
   )
 }
 
-export default ConfigHelper;
+const IntlConfigHelper = ()=>{
+  return (
+    <LocaleProvider>
+      <ConfigHelper/>
+    </LocaleProvider>
+  )
+}
+
+export default IntlConfigHelper;