+++ extensions/iceworks-project-creator/web/src/pages/CreateProject/components/ScaffoldMarket/index.tsx
@@ -9,10 +9,12 @@ import { IMaterialSource, IMaterialScaffold } from '@iceworks/material-utils';
 import { mainScaffoldsList, tsScaffoldsList, jsScaffoldsList } from '@/constant';
 import { IScaffoldMarket } from '@/types';
 import styles from './index.module.scss';
+import { useIntl } from 'react-intl';
 
 const projectTypes = ['react', 'rax', 'vue'];
 
 const ScaffoldMarket = ({ onScaffoldSelect, curProjectField, children, onOpenConfigPanel, materialSources }) => {
+  const intl = useIntl();
   const [selectedSource, setSelectedSource] = useState<any>({});
   const [mainScaffolds, setMainScaffolds] = useState<IMaterialScaffold[]>([]);
   const [otherScaffolds, setOtherScaffolds] = useState<IMaterialScaffold[]>([]);
@@ -104,7 +106,7 @@ const ScaffoldMarket = ({ onScaffoldSelect, curProjectField, children, onOpenCon
                 }
                 content={item.description}
                 selected={selectedSource.name && selectedSource.name === item.name}
-                style={{ width: 160, height: 110 }}
+                style={{ width: 160, height: 120 }}
                 onClick={() => onMaterialSourceClick(item)}
               />
             })}
@@ -118,6 +120,8 @@ const ScaffoldMarket = ({ onScaffoldSelect, curProjectField, children, onOpenCon
           {loading ? <Loading visible={loading} className={styles.loading} /> : <>
             <div className={styles.mainScaffolds}>
               {!!mainScaffolds.length ? mainScaffolds.map(item => {
+                // tsScaffoldsList and jsScaffoldsList only contain the official scaffolds
+                // so the TypeScript and JavaScript logo only display in official scaffolds
                 const scaffoldType = tsScaffoldsList.includes(item.source.npm) ? 'ts' :
                   jsScaffoldsList.includes(item.source.npm) ? 'js' :
                     '';
@@ -129,7 +133,7 @@ const ScaffoldMarket = ({ onScaffoldSelect, curProjectField, children, onOpenCon
                     title={
                       <div className={styles.cardTitle}>
                         {scaffoldType && <img src={require(`@/assets/${scaffoldType}.svg`)} alt="languageType" width={20} height={20} />}
-                        <div>{item.title.replace(' - TS', '')}</div>
+                        <div>{scaffoldType ? item.title.replace(' - TS', '').replace(' - JS', '') : item.title}</div>
                       </div>
                     }
                     content={item.description}
@@ -139,13 +143,15 @@ const ScaffoldMarket = ({ onScaffoldSelect, curProjectField, children, onOpenCon
                   />
                 )
               }) :
-                <NotFound description="暂无模板" />
+                <NotFound description={intl.formatMessage({ id: 'web.iceworksProjectCreator.ScaffoldMarket.noTemplate' })} />
               }
             </div>
             {!!otherScaffolds.length && <Collapse className={styles.collapse}>
-              <Collapse.Panel title="查看更多">
+              <Collapse.Panel title={intl.formatMessage({ id: 'web.iceworksProjectCreator.ScaffoldMarket.more' })}>
                 <div className={styles.collapseScaffolds}>
                   {otherScaffolds.map(item => {
+                    // tsScaffoldsList and jsScaffoldsList only contain the official scaffolds
+                    // so the TypeScript and JavaScript logo only display in official scaffolds
                     const scaffoldType = tsScaffoldsList.includes(item.source.npm) ? 'ts' :
                       jsScaffoldsList.includes(item.source.npm) ? 'js' :
                         '';
@@ -157,7 +163,7 @@ const ScaffoldMarket = ({ onScaffoldSelect, curProjectField, children, onOpenCon
                         title={
                           <div className={styles.cardTitle}>
                             {scaffoldType && <img src={require(`@/assets/${scaffoldType}.svg`)} alt="languageType" width={20} height={20} />}
-                            <div>{item.title.replace(' - JS', '')}</div>
+                            <div>{scaffoldType ? item.title.replace(' - JS', '').replace(' - TS', '') : item.title}</div>
                           </div>
                         }
                         content={item.description}