+++ packages/generate-project/src/index.ts
@@ -2,6 +2,8 @@ import * as ora from 'ora';
 import {
   isAliNpm, getNpmTarball, getAndExtractTarball
 } from 'ice-npm-utils';
+import { ALI_NPM_REGISTRY } from '@iceworks/constant';
+import ejsRenderDir from './ejsRenderDir';
 import formatProject from './fommatProject';
 import checkEmpty from './checkEmpty';
 
@@ -10,10 +12,30 @@ export {
   checkEmpty,
 };
 
+export interface IEjsOptions {
+  targets?: string[];
+  miniappType?: 'runtime' | 'compile';
+  mpa?: boolean;
+};
+
 export async function downloadAndGenerateProject(
-  projectDir: string, npmName: string, version?: string, registry?: string, projectName?: string
+  projectDir: string,
+  npmName: string,
+  version?: string,
+  registry?: string,
+  projectName?: string,
+  ejsOptions?: IEjsOptions,
 ): Promise<void> {
   registry = registry || await getNpmRegistry(npmName);
+
+  // 根据模板创建项目支持的参数
+  ejsOptions = {
+    targets: ['web'],
+    miniappType: 'runtime',
+    mpa: false,
+    ...ejsOptions
+  };
+
   let tarballURL: string;
   try {
     tarballURL = await getNpmTarball(npmName, version || 'latest', registry);
@@ -25,7 +47,7 @@ export async function downloadAndGenerateProject(
     }
   }
 
-  console.debug('download tarballURL', tarballURL);
+  console.log('download tarballURL', tarballURL);
 
   const spinner = ora('download npm tarball start').start();
   await getAndExtractTarball(
@@ -34,14 +56,16 @@ export async function downloadAndGenerateProject(
     (state) => {
       spinner.text = `download npm tarball progress: ${Math.floor(state.percent * 100)}%`;
     },
-    formatFilename,
+    formatFilename
   );
   spinner.succeed('download npm tarball successfully.');
+
+  await ejsRenderDir(projectDir, ejsOptions);
 
   try {
     await formatProject(projectDir, projectName);
   } catch (err) {
-    console.warn('formatProject error');
+    console.warn('formatProject error', err.message);
   }
 };
 
@@ -49,7 +73,7 @@ async function getNpmRegistry(npmName: string): Promise<string> {
   if (process.env.REGISTRY) {
     return process.env.REGISTRY;
   } else if (isAliNpm(npmName)) {
-    return 'https://registry.npm.alibaba-inc.com';
+    return ALI_NPM_REGISTRY;
   } else {
     return 'https://registry.npm.taobao.org';
   }