+++ packages/ice-npm-utils/src/__tests__/index.ts
@@ -1,6 +1,8 @@
 import 'jest-extended';
 import * as path from 'path';
 import { tmpdir } from 'os';
+import * as rimraf from 'rimraf';
+import { ALI_NPM_REGISTRY, ALI_UNPKG_URL } from '@iceworks/constant';
 import {
   getNpmRegistry,
   getUnpkgHost,
@@ -13,20 +15,16 @@ import {
   getAndExtractTarball,
 } from '../index';
 
-import rimraf = require('rimraf');
-
 const defaultRegistry = 'https://registry.npm.taobao.org';
 
 jest.setTimeout(10 * 1000);
 
 test('getNpmRegistry', () => {
-  const aliRegistry = 'https://registry.npm.alibaba-inc.com';
-
   expect(getNpmRegistry('koa')).toBe(defaultRegistry);
   expect(getNpmRegistry('@alixxx/ice-test')).toBe(defaultRegistry);
-  expect(getNpmRegistry('@ali/ice-test')).toBe(aliRegistry);
-  expect(getNpmRegistry('@alife/ice-test')).toBe(aliRegistry);
-  expect(getNpmRegistry('@alipay/ice-test')).toBe(aliRegistry);
+  expect(getNpmRegistry('@ali/ice-test')).toBe(ALI_NPM_REGISTRY);
+  expect(getNpmRegistry('@alife/ice-test')).toBe(ALI_NPM_REGISTRY);
+  expect(getNpmRegistry('@alipay/ice-test')).toBe(ALI_NPM_REGISTRY);
 });
 
 
@@ -43,13 +41,12 @@ test('getUnpkgHost custom host', () => {
 
 test('getUnpkgHost', () => {
   const defaultRegistry = 'https://unpkg.com';
-  const aliRegistry = 'https://unpkg.alibaba-inc.com';
 
   expect(getUnpkgHost('koa')).toBe(defaultRegistry);
-  expect(getUnpkgHost('@ali/ice-test')).toBe(aliRegistry);
-  expect(getUnpkgHost('@alife/ice-test')).toBe(aliRegistry);
-  expect(getUnpkgHost('@alipay/ice-test')).toBe(aliRegistry);
-  expect(getUnpkgHost('@kaola/ice-test')).toBe(aliRegistry);
+  expect(getUnpkgHost('@ali/ice-test')).toBe(ALI_UNPKG_URL);
+  expect(getUnpkgHost('@alife/ice-test')).toBe(ALI_UNPKG_URL);
+  expect(getUnpkgHost('@alipay/ice-test')).toBe(ALI_UNPKG_URL);
+  expect(getUnpkgHost('@kaola/ice-test')).toBe(ALI_UNPKG_URL);
   expect(getUnpkgHost('@alixxx/ice-test')).toBe(defaultRegistry);
 });