+++ extensions/iceworks-component-builder/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.2.0
+
+- feat: support en-US 
+
+## 0.1.10
+
+Change the material setting entry to Iceworks setting. 
+
 ## 0.1.9
 
 Use webpack to build this extension.